+++ scripts/test/CustomerUser.t
@@ -23,7 +23,7 @@ $ConfigObject->Set(
     Value => 0,
 );
 
-my $DatabaseCaseSensitive = $Self->{DBObject}->{Backend}->{'DB::CaseSensitive'};
+my $DatabaseCaseSensitive                = $Self->{DBObject}->{Backend}->{'DB::CaseSensitive'};
 my $CustomerDatabaseCaseSensitiveDefault = $ConfigObject->{CustomerUser}->{Params}->{CaseSensitive};
 
 my $CustomerUserObject = Kernel::System::CustomerUser->new(
@@ -180,12 +180,13 @@ for my $Key ( 1 .. 3, 'ä', 'カス' ) {
         ConfigObject => $ConfigObject,
     );
     $CacheObject->CleanUp();
+
     # Customer Search
     %List = $CustomerUserObject->CustomerSearch(
-        Search => lc($UserRand . '-Customer-Update-Id'),
-        ValidID    => 1,
+        Search  => lc( $UserRand . '-Customer-Update-Id' ),
+        ValidID => 1,
     );
-    
+
     if ($DatabaseCaseSensitive) {
 
         $Self->False(
@@ -202,7 +203,7 @@ for my $Key ( 1 .. 3, 'ä', 'カス' ) {
 
     # CustomerIDList
     my @List = $CustomerUserObject->CustomerIDList(
-        SearchTerm => lc($UserRand . '-Customer-Update-Id'),
+        SearchTerm => lc( $UserRand . '-Customer-Update-Id' ),
         ValidID    => 1,
     );
 
@@ -231,8 +232,8 @@ for my $Key ( 1 .. 3, 'ä', 'カス' ) {
 
     # Customer Search
     %List = $CustomerUserObject->CustomerSearch(
-        Search => lc($UserRand . '-Customer-Update-Id'),
-        ValidID    => 1,
+        Search  => lc( $UserRand . '-Customer-Update-Id' ),
+        ValidID => 1,
     );
     $Self->True(
         $List{$UserID},
@@ -241,7 +242,7 @@ for my $Key ( 1 .. 3, 'ä', 'カス' ) {
 
     # CustomerIDList
     @List = $CustomerUserObject->CustomerIDList(
-        SearchTerm => lc($UserRand . '-Customer-Update-Id'),
+        SearchTerm => lc( $UserRand . '-Customer-Update-Id' ),
         ValidID    => 1,
     );
 
@@ -251,7 +252,9 @@ for my $Key ( 1 .. 3, 'ä', 'カス' ) {
         "CustomerIDList() - no SearchTerm - $UserID (CaseSensitive = 0)",
     );
 
-    $ConfigObject->{CustomerUser}->{Params}->{CaseSensitive} = $CustomerDatabaseCaseSensitiveDefault;
+    $ConfigObject->{CustomerUser}->{Params}->{CaseSensitive}
+        = $CustomerDatabaseCaseSensitiveDefault;
+
     # END CaseSensitive
 
     @List = $CustomerUserObject->CustomerIDList(